+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/notification/api/domain/NotificationLevel.java
@@ -1,5 +1,7 @@
 package eu.bcvsolutions.idm.core.notification.api.domain;
 
+import org.springframework.http.HttpStatus;
+
 /**
  * Notification level
  * 
@@ -10,5 +12,21 @@ public enum NotificationLevel {
 	SUCCESS,
 	INFO,
 	WARNING,
-	ERROR
+	ERROR;
+	
+	/**
+	 * Returns level for given http status
+	 *  
+	 * @param status
+	 * @return
+	 */
+	public static NotificationLevel getLevel(HttpStatus status) {
+		if (status.is5xxServerError()) {
+            return NotificationLevel.ERROR;
+        } 
+		if (status.is2xxSuccessful()) {
+            return NotificationLevel.SUCCESS;
+        }
+        return NotificationLevel.WARNING;
+	}
 }