+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/notification/service/impl/DefaultIdmNotificationConfigurationService.java
@@ -26,7 +26,7 @@ import eu.bcvsolutions.idm.core.api.service.AbstractReadWriteDtoService;
 import eu.bcvsolutions.idm.core.api.service.ModuleService;
 import eu.bcvsolutions.idm.core.notification.api.domain.NotificationLevel;
 import eu.bcvsolutions.idm.core.notification.api.dto.BaseNotification;
-import eu.bcvsolutions.idm.core.notification.api.dto.NotificationConfigurationDto;
+import eu.bcvsolutions.idm.core.notification.api.dto.IdmNotificationConfigurationDto;
 import eu.bcvsolutions.idm.core.notification.api.service.IdmNotificationConfigurationService;
 import eu.bcvsolutions.idm.core.notification.api.service.NotificationSender;
 import eu.bcvsolutions.idm.core.notification.entity.IdmNotificationConfiguration;
@@ -42,7 +42,7 @@ import eu.bcvsolutions.idm.core.security.api.domain.BasePermission;
  */
 @Service("notificationConfigurationService")
 public class DefaultIdmNotificationConfigurationService 
-		extends AbstractReadWriteDtoService<NotificationConfigurationDto, IdmNotificationConfiguration, EmptyFilter>
+		extends AbstractReadWriteDtoService<IdmNotificationConfigurationDto, IdmNotificationConfiguration, EmptyFilter>
 		implements IdmNotificationConfigurationService {
 	
 	private final IdmNotificationConfigurationRepository repository;
@@ -66,7 +66,7 @@ public class DefaultIdmNotificationConfigurationService
 	
 	@Override
 	@Transactional
-	public NotificationConfigurationDto save(NotificationConfigurationDto dto, BasePermission... permission) {
+	public IdmNotificationConfigurationDto save(IdmNotificationConfigurationDto dto, BasePermission... permission) {
 		Assert.notNull(dto);
 		//
 		// check duplicity
@@ -91,7 +91,7 @@ public class DefaultIdmNotificationConfigurationService
 				if (topicToCreate.contains(topic) || count == 0) {
 					topicToCreate.add(topic);
 					UUID template = config.getTemplate();
-					NotificationConfigurationDto notConfiguration = new NotificationConfigurationDto(config);
+					IdmNotificationConfigurationDto notConfiguration = new IdmNotificationConfigurationDto(config);
 					notConfiguration.setTemplate(template);
 					repository.save(toEntity(notConfiguration, null));
 				}
@@ -162,12 +162,12 @@ public class DefaultIdmNotificationConfigurationService
 	}
 
 	@Override
-	public NotificationConfigurationDto getConfigurationByTopicLevelNotificationType(String topic, NotificationLevel level, String notificationType) {
+	public IdmNotificationConfigurationDto getConfigurationByTopicLevelNotificationType(String topic, NotificationLevel level, String notificationType) {
 		return toDto(this.repository.findByTopicAndLevelAndNotificationType(topic, level, notificationType));
 	}
 
 	@Override
-	public List<NotificationConfigurationDto> getConfigurations(String topic, NotificationLevel level) {
+	public List<IdmNotificationConfigurationDto> getConfigurations(String topic, NotificationLevel level) {
 		return toDtos(repository.findByTopicAndLevel(topic, level), false);
 	}